Added a lot of the repository pattern + base for skills + inventory
A character can now also retrieve the weapons that character has.

I've also added a lot of other repositories + contexts so I can start adding more feature to my logic layer soon.

To do for when I have time tomorrow: Add armour and other items to the inventory of a character as well.